requirements.txt
pdfplumber
python-docx
spacy
scikit-learn
pandas
numpy
